Implement command buffers, Implement ModelDisplay for Observer window, Fix cubes.svg being off center

+++ Engine/Mesh.hpp
@@ -25,8 +25,8 @@ struct Mesh {
     std::string Name{};
     std::optional<VertexList> Vertices{};
     std::optional<IndexList> Indices{};
-    size_t VertexCount{};
-    size_t IndexCount{};
+    unsigned VertexCount{};
+    unsigned IndexCount{};
 
     glm::vec3 Min{}; // Element-wise min position. May not correspond to a real vertex in the Mesh.
     glm::vec3 Max{}; // Element-wise max position. May not correspond to a real vertex in the Mesh.
